hisi_riscv_rt::interrupt

Function cause

pub fn cause() -> Trap<Interrupt, Exception>

Retrieves the cause of a trap in the current hart (machine mode).

If the raw cause is not a valid interrupt or exception for the target, it panics.