hisi_riscv_rt

Module prelude

Prelude: commonly used runtime types.

Enums

Interrupt

External interrupts. These interrupts are handled by the external peripherals.

Attribute Macros

entry

Attribute to declare the entry point of the program

pre_init

Attribute to mark which function will be called at the beginning of the reset handler.